Analyte Object

Analyte class is an object that gathers information about one specific Analyte on the Luminex plate

Public fields

id

A numeric value that represents the unique identifier of the analyte

analyte_name

A character value that represents the name of the analyte

bead_count

? not exactly sure what this field means tbh

analysis_type

Type of the analyte

units

Units of the analyte in which the results are expressed

Methods

Public methods

Analyte$new()

Analyte$print()

Analyte$join()

Analyte$clone()

Method new()

Creates a new instance of the Analyte class

Usage

Analyte$new(
  id,
  analyte_name,
  bead_count = NA,
  analysis_type = NULL,
  units = NULL
)

Arguments

id

(numeric(1))
A numeric value that represents the unique identifier of the analyte

analyte_name

(character(1))
A character value that represents the name of the analyte

bead_count

(numeric(1))

analysis_type

(character(1))
Type of the analyte

units

(character(1))
Units of the analyte in which the results are expressed

Method print()

Prints the information about the analyte

Usage

Analyte$print(...)

Method join()

Joins the data of two analytes Performs checks to ensure that the analytes can be joined If the analytes can be joined, the bead_count, analysis_type and units are updated in the current analyte object

Usage

Analyte$join(new_analyte)

Arguments

new_analyte

(Analyte)
The analyte object to be joined with the current analyte object - this is rather a technical inner functionand should rarely be used by the user

Method clone()

The objects of this class are cloneable with this method.

Usage

Analyte$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

etramp <- Analyte$new(id = 73, analyte_name = "Etramp5_ag1", bead_count = 50)
print(etramp)
#> Analyte:  Etramp5_ag1 
#> ID:  73 
#> Bead Count:  50 
#> Analysis Type: