Sample

A class to represent the sample. It contains all the necessary information about the sample

Public fields

id

Sample id, when reading from a plate it is a id extracted from the sample position on the plate

sample_name

Sample name

sample_type

A sample type - is the sample used for testing, is it blank, positive or negative control

sample_location

Position on the Plate

warnings

list of warnings warnings are read from the file or generated during the reading process

errors

lit of errors; errors are read from the file or generated during the reading process

data

The dataframe containing informations about the analytes within the sample. The dataframe rows are different measures, such as Median, Net MFI, etc. whereas the columns represent different analytes

Methods

Public methods

Sample$new()

Sample$print()

Sample$join()

Sample$clone()

Method new()

Initializes the sample object

Usage

Sample$new(
  id,
  sample_name,
  sample_type = NULL,
  sample_location = NULL,
  warnings = list(),
  errors = list(),
  data = data.frame()
)

Arguments

id

(numeric(1))
The id of the sample - usually just the sample row number, can be derived from the sample location

sample_name

(character(1))
The name of the sample

sample_type

(SampleType)
The type of the sample

The location of the sample on the plate

warnings

(list)
List containing all the warnings connected to the sample - generated by the validity checks or read from file

errors

(list)
List containing all the critical errors connected to the sample - generated by the validity checks or read from file

data

(data.frame)
The dataframe containing the informations about analytes within the sample

Method print()

Prints the sample information

Usage

Sample$print(...)

Method join()

Joins the data of two samples

Usage

Sample$join(new_sample)

Method clone()

The objects of this class are cloneable with this method.

Usage

Sample$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# TODO