SampleLocation Object

SampleLocation defines a location of a sample on the Luminex plate The location is defined by the row and column number and explicitly states the location of the sample

Public fields

col

column number of sample on the Luminex plate

row

row number of sample on the Luminex plate

Active bindings

row_letter

row where sample is located as a letter. For instance 1 -> A

location_name

Lcation of the sample formatted in readable way - (row_letter, column)

Methods

Public methods

SampleLocation$new()

SampleLocation$print()

SampleLocation$join()

SampleLocation$clone()

Method new()

Creates a new instance of the SampleLocation class

Usage

SampleLocation$new(row = NA, col = NA)

Arguments

row

(numeric(1))
A numeric value that represents the row number of the sample location

col

(numeric(1))
A numeric value that represents the column number of the sample location

Method print()

Prints the information about the sample location

Usage

SampleLocation$print(...)

Method join()

Joins the data of two sample locations Performs checks to ensure that the sample locations can be joined Essentially checks if the two sample locations are identical or at least one of them is empty

Usage

SampleLocation$join(new_location)

Method clone()

The objects of this class are cloneable with this method.

Usage

SampleLocation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

sample_location <- SampleLocation$new(col = 1, row = 4)

sample_location$location_name
#> [1] "(D, 1)"

sample_location <- SampleLocation$parse_sample_location("65(1,F5)")
sample_location$location_name
#> [1] "(F, 5)"