SampleType Object

SampleType defines the type of the sample and the dilution factor

Public fields

sample_type

A type of the sample. The possible sample types are:

("BLANK")
- background sample used for testing

("STANDARD CURVE")
- a sample used to create a standard curve

("POSITIVE CONTROL")
- a positive control sample used for testing

("NEGATIVE CONTROL")
- a negative control sample used for testing

("TEST")
- the actual sample used for testing

dilution_factor

A numeric value that represents the dilution factor of the sample. Used only in the case of control samples

Active bindings

character_dilution_factor

the dilution factor in the format 1/x - useful for printing

Methods

Public methods

SampleType$new()

SampleType$print()

SampleType$join()

SampleType$clone()

Method new()

Creates a new instance of the SampleType class

Usage

SampleType$new(sample_type, dilution_factor = NA, validate_dilution = TRUE)

Arguments

sample_type

(character(1))
A character value that represents the type of the sample Possible values are: "BLANK", "POSITIVE CONTROL", "NEGATIVE CONTROL", "TEST".

dilution_factor

(numeric(1))
A numeric value that represents the dilution factor of the sample. Used only in the case of control samples

Whether

to validate the object for the dilution factor. Used in the reading scripts to dynamically assign values. Default to TRUE

Method print()

Prints the sample type and the dilution factor

Usage

SampleType$print(...)

Method join()

Joins the data of two sample types Performs checks to ensure that the sample types can be joined Essentially checks if the two sample types are identical or at least one of them is empty

Usage

SampleType$join(new_sample_type)

Method clone()

The objects of this class are cloneable with this method.

Usage

SampleType$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.