Plot standard curve of a certain analyte with fitted model

Function plots the values of standard curve samples and the fitted model.

plot_standard_curve_analyte_with_model(
  plate,
  model,
  data_type = "Median",
  decreasing_dilution_order = TRUE,
  log_scale = c("all"),
  plot_asymptote = TRUE,
  plot_test_predictions = TRUE,
  plot_blank_mean = TRUE,
  plot_dilution_bounds = TRUE,
  verbose = TRUE
)

Arguments

plate

Plate object

model

fitted Model object, which predictions we want to plot

data_type

Data type of the value we want to plot - the same datatype as in the plate file. By default equals to Median

decreasing_dilution_order

If TRUE the dilutions are plotted in decreasing order, TRUE by default.

log_scale

Which elements on the plot should be displayed in log scale. By default "all". If NULL or c() no log scale is used, if "all" or c("dilutions", "MFI") all elements are displayed in log scale.

plot_asymptote

If TRUE the asymptotes are plotted, TRUE by default

plot_test_predictions

If TRUE the predictions for the test samples are plotted, TRUE by default The predictions are obtained through extrapolation of the model

plot_blank_mean

If TRUE the mean of the blank samples is plotted, TRUE by default

plot_dilution_bounds

If TRUE the dilution bounds are plotted, TRUE by default

verbose

If TRUE prints messages, TRUE by default

Value

a ggplot object with the plot