Process a plate and save computed dilutions to a CSV

The behavior can be summarized as follows:

Adjust blanks if not already done.

Fit a model to each analyte using standard curve samples.

Compute dilutions for each analyte using the corresponding model.

Aggregate computed dilutions into a single data frame.

Save the computed dilutions to a CSV file.

process_plate(
  plate,
  output_path = NULL,
  data_type = "Median",
  adjust_blanks = FALSE,
  verbose = TRUE,
  ...
)

Arguments

plate

(Plate()) a plate object

output_path

(character(1)) path to save the computed dilutions. If not provided the file will be saved in the working directory with the name dilutions_{plate_name}.csv. Where the {plate_name} is the name of the plate.

data_type

(character(1)) type of data to use for the computation. Median is the default

adjust_blanks

(logical(1)) adjust blanks before computing dilutions. Default is FALSE

verbose

(logical(1)) print additional information. Default is TRUE

...

Additional arguments to be passed to the fit model function (create_standard_curve_model_analyte)

Examples

plate_file <- system.file("extdata", "CovidOISExPONTENT.csv", package = "PvSTATEM")
layout_file <- system.file("extdata", "CovidOISExPONTENT_layout.csv", package = "PvSTATEM")

plate <- read_luminex_data(plate_file, layout_file)
#> Reading Luminex data from: /home/runner/work/_temp/Library/PvSTATEM/extdata/CovidOISExPONTENT.csv
#> using format xPONENT
#> 
#> New plate object has been created with name: CovidOISExPONTENT!
#> 
process_plate(plate) # create and save dataframe with computed dilutions
#> Fitting the models and computing the dilutions for each analyte
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Warning: One (or more) of the values were greater or equal to the estimated top asymptote.
#> These values have been replaced by the maximal possible value the model can estimate.
#> Saving the computed dilutions to a CSV file located in: 'dilutions_CovidOISExPONTENT.csv'