Plate

A class to represent the luminex plate. It contains information about the samples and analytes that were examined on the plate as well as some additional metadata and batch info

Value

sample object of given sample name or id

analyte id of given name

Public fields

analytes

list of all analytes measured during an examination

samples

list of all samples examined on a plate

batch_info

a raw list containing all the metadata about the plate read from the Luminex file

calibration_info

a list containing calibration logs read from Luminex file

audit_logs

a list containing audit logs read from Luminex file

plate_name

plate name obtained from filename

Active bindings

number_of_samples

number of samples stored in the current plate

analyte_names

list of all analyte names saved in the plate

sample_names

list of all sample names

examination_date

Metadata: date of the examination

batch_name

Metdata: batch name

min_events_per_bead

minimal number of events that is valid for one bead - sample and analyte

check_if_blanks_already_adjusted

flag that specifies if the blanks were already adjusted and its MFI values subtracted from remaining samples

warnings

list of lists of all warnings from all samples

errors

list of lists of all errors from all samples

Methods

Public methods

Plate$new()

Plate$print()

Plate$summary()

Plate$add_results_block()

Plate$check_batch_info()

Plate$check_analyte_consistency()

Plate$check_beads_number()

Plate$get_sample()

Plate$get_sample_by_type()

Plate$get_analyte_id()

Plate$get()

Plate$blank_adjustment()

Plate$check_MFI_after_adjustment()

Plate$copy()

Plate$clone()

Method new()

creates a new Plate object

Usage

Plate$new(
  analytes = list(),
  samples = list(),
  batch_info = list(),
  calibration_info = list(),
  audit_logs = list(),
  plate_name = ""
)

Arguments

analytes

(list) list of all analytes measured during an examination

samples

(list) list of all samples examined on a plate

batch_info

(list)
a raw list containing all the metadata about the plate read from the Luminex file

calibration_info

(list)
a list containing calibration logs read from Luminex file

audit_logs

(list) a list containing audit logs read from Luminex file

Method print()

Usage

Plate$print(...)

Method summary()

Function outputs basic information about the plate, such as examination date, batch name, and sample types

Usage

Plate$summary(..., include_names = FALSE)

Arguments

include_names

If include_names parameter is TRUE, apart from count of control samples, provides also their names.By default FALSE

Method add_results_block()

function adds block of information to the current plate

the analysis type should be added after the analysis, otherwise there is no enough information about samples

Usage

Plate$add_results_block(data_type, parsed_block)

Method check_batch_info()

Function verifies if the batch info contain valid information

Usage

Plate$check_batch_info()

Method check_analyte_consistency()

checks analyte consistency - verifies if all of the analytes contained within the samples are listed in the analytes list of the plate object

Usage

Plate$check_analyte_consistency()

Method check_beads_number()

function performs validity check - verifies which samples and analytes saved in the plate have the bead count lower than min_events_per_bead

Usage

Plate$check_beads_number(min_events_per_bead = self$min_events_per_bead)

Arguments

min_events_per_bead

lower bound of acceptable number of events. By default equals to min_events_per_bead parameter saved in the plate object

Method get_sample()

Usage

Plate$get_sample(sample)

Arguments

sample

sample name or its id

Method get_sample_by_type()

Function returns list of samples filtered by the type

Usage

Plate$get_sample_by_type(sample_type, exclude = FALSE)

Arguments

sample_type

type of the sample to be filtered. Possible values are: SampleType$valid_sample_types

exclude

If FALSE returns list of samples with given sample_type, otherwise returns all samples except for the specified sample_type

Method get_analyte_id()

Usage

Plate$get_analyte_id(analyte_name)

Method get()

Function returns data for a specific analyte and sample.

Usage

Plate$get(analyte, sample, data_type = NULL)

Arguments

analyte

An analyte name or its id of which data we want to extract

sample

sample name or id

data_type

if NULL returns whole column of the dataframe containing information about the sample. Otherwise tries to select the exact data_type from the dataframe and returns a single value

Returns

Data about a sample and analyte

Method blank_adjustment()

Function adjusts the values of test samples substracting values from BLANK samples to remove background light In short it substracts the values from data in all samples, except from Blanks. It does not substract values from Count values

Usage

Plate$blank_adjustment(method = "avg", inplace = "TRUE")

Arguments

method

How the values of different blanks should be aggregated. By default avg. For now it is the only available method

inplace

Whether the method should produce new plate with adjusted values or not, By default TRUE - operates on the current plate.

Method check_MFI_after_adjustment()

Function verifies if there are any MFI values below zero after blank removal

Usage

Plate$check_MFI_after_adjustment()

Method copy()

performs copy of the plate

Usage

Plate$copy()

Method clone()

The objects of this class are cloneable with this method.

Usage

Plate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# TODO